+++ qiskit/dagcircuit/dagnode.py
